Agrego de permisos a los controladores

+++ app/Http/Controllers/MenuController.php
@@ -9,14 +9,20 @@ use App\Http\Requests\UpdateMenuRequest;
 use App\Http\Resources\MenuResource;
 use Carbon\Carbon;
 use Illuminate\Support\Arr;
+use Exception;
+use Illuminate\Support\Facades\Auth;
 
 class MenuController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
+    
     public function index()
     {
+        if (!Auth::guard('api')->user()->can('view_menu')) {
+            return response()->json([
+                'status' => 403,
+                'message' => 'No tiene permiso para visualizar el menu.'
+            ], 403);
+        }
         try {
             $today = Carbon::now();
             $menu = Menu::whereDate('date_menu', $today)->get();
@@ -31,7 +37,7 @@ class MenuController extends Controller
                     'menus' => $menu->toArray() 
                 ]);
             }
-        } catch (\Exception $e) {
+        } catch (Exception $e) {
             return response()->json([
                     'status' => 404,
                     'message' => 'No se encontraron registros' . $e->getMessage()  
@@ -39,46 +45,66 @@ class MenuController extends Controller
         }
 
     }
-    /**
-     * Store a newly created resource in storage.
-     */
+
     public function store(StoreMenuRequest $request)
     {
+        if (!Auth::guard('api')->user()->can('create_menu')) {
+            return response()->json([
+                'status' => 403,
+                'message' => 'No tiene permiso para guardar menu.'
+            ], 403);
+        }
         return new MenuResource(Menu::create($request->all()));
     }
 
     public function BlukStore(BlukStoreMenuRequest $request){
-        
-        $dataCheks = Carbon::now()->toDateString();
-        $menusExists = Menu::where('date_menu', $dataCheks)->exists();
-
-        if ($menusExists) {
+        if (!Auth::guard('api')->user()->can('create_menu_bluk')) {
             return response()->json([
-                'status' => 409,
-                'message' => 'El MENU ya fue cargado'
-            ], 409);
-        }else {
-            $bluk = collect($request->all())->map(function ($arr, $key){
-                return Arr::except($arr, ['foodCategory', 'ingredient', 'dateMenu']);
-            });
-            Menu::insert($bluk->toArray());
-            return response()->json([
-                'status' => 200,
-                'message' => "Menu guardado Correctamente"
-            ], 200);
+                'status' => 403,
+                'message' => 'No tiene permiso de cargar menu.'
+            ], 403);
         }
-        
-       
+        try {
+            $dataCheks = Carbon::now()->toDateString();
+            $menusExists = Menu::where('date_menu', $dataCheks)->exists();
+
+            if ($menusExists) {
+                return response()->json([
+                    'status' => 409,
+                    'message' => 'El MENU ya fue cargado'
+                ], 409);
+            }else {
+                $bluk = collect($request->all())->map(function ($arr, $key){
+                    return Arr::except($arr, ['foodCategory', 'ingredient', 'dateMenu']);
+                });
+                Menu::insert($bluk->toArray());
+                return response()->json([
+                    'status' => 200,
+                    'message' => "Menu guardado Correctamente"
+                ], 200);
+            }
+        } catch (Exception $e) {
+            return response()->json([
+                'status' => 500,
+                'message' => 'Error en registrar menu: '. $e->getMessage()
+            ]);
+        }
     }
     public function show($date)
     {
+        if (!Auth::guard('api')->user()->can('show_menu')) {
+            return response()->json([
+                'status' => 403,
+                'message' => 'No tiene permiso para visualizar el menu.'
+            ], 403);
+        }
         try {
             $menu = Menu::whereDate('date_menu', $date)->get();
             return response()->json([
                 'status' => 200,
                 $menu
             ], 200);
-        } catch (\Exception $e) {
+        } catch (Exception $e) {
             return response()->json([
                 'status' => 401,
                 'message' => 'Error en encontrar Menu' . $e->getMessage() 
@@ -86,11 +112,14 @@ class MenuController extends Controller
         }
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
     public function update(UpdateMenuRequest $request, $id)
     {
+        if (!Auth::guard('api')->user()->can('update_menu')) {
+            return response()->json([
+                'status' => 403,
+                'message' => 'No tiene permiso para editar el menu.'
+            ], 403);
+        }
         try {
             $menu = Menu::where('id_menu', $id)->firstOrFail();
             $validated = $request->validated();
@@ -103,7 +132,7 @@ class MenuController extends Controller
                 'status' => 200,
                 'message' => 'Se actualizo el item correctamente'
             ], 200);
-        } catch (\Exception $e) {
+        } catch (Exception $e) {
             return response()->json([
                 'status' => 409,
                 'message' => 'Error al actualizar el item',
@@ -114,6 +143,12 @@ class MenuController extends Controller
 
     public function destroy($date)
     {
+        if (!Auth::guard('api')->user()->can('delete_menu')) {
+            return response()->json([
+                'status' => 403,
+                'message' => 'No tiene permiso para eliminar el menu.'
+            ], 403);
+        }
         try {
             Menu::whereDate('date_menu', $date)->delete();
             return response()->json([
@@ -121,7 +156,7 @@ class MenuController extends Controller
                 'message' => 'Menu eliminado Correctamente'
             ], 200);
 
-        } catch (\Exception $e) {
+        } catch (Exception $e) {
             return response()->json([
                 'status' => 401,
                 'message' => 'Error al eliminar Menu'. $e->getMessage()
